backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
openai==1.51.2
Pillow==10.4.0
httpx==0.27.2
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy==2.1.3
qrcode==7.4.2
